requirements.txt
azure-storage-blob==12.11.0
azure-identity==1.10.0
retry==0.9.2
